+++ samples/MinimalApi/Program.cs
@@ -14,8 +14,9 @@ services.AddSwaggerGen();
 // Add FV validators
 services.AddValidatorsFromAssemblyContaining<Program>();
 
-// Add FV Rules to swagger
-services.AddFluentValidationRulesToSwagger();
+// Add FV Rules to swagger (ADR-007: this sample exercises the document-filter pipeline)
+services.AddFluentValidationRulesToSwagger(
+    configureRegistration: options => options.UseDocumentFilter = true);
 
 var app = builder.Build();